rpl: split mod into multiple files
